// src/generated/parser/clothing-data.ts
export interface ClothingSlotAndColor {
  Slot: string;
  color: UnityColor;
}

export interface ClothingSlotsAndColours {
  SlotAndColours: ClothingSlotAndColor[];
  AllColor: UnityColor;
}

export interface ClothingMealEffect {
  MealEffectType: number;
  Chance: number;
}

export interface ClothingMonoBehaviourHeader {
  m_GameObject: UnityPPtr;
  m_Enabled: boolean;
  m_Script: UnityPPtr;
  m_Name: string;
}

export interface ClothingDataFile {
  header: ClothingMonoBehaviourHeader;
  SkeletonData: UnityPPtr;
  ClothingType: number;
  ProtectionType: number;
  ForSale: boolean;
  SpecialClothing: boolean;
  IsSecret: boolean;
  IsDLC: boolean;
  IsMajorDLC: boolean;
  IsHeretic: boolean;
  IsCultist: boolean;
  IsSinful: boolean;
  IsPilgrim: boolean;
  HideOnTailorMenu: boolean;
  CanBeCrafted: boolean;
  Effects: ClothingMealEffect[];
  Variants: string[];
  SlotAndColours: ClothingSlotsAndColours[];
}