// src/generated/parser/worshipper-data.ts
export enum WorshipperDropLocation {
  Dungeon1 = 0,
  Dungeon2 = 1,
  Dungeon3 = 2,
  Dungeon4 = 3,
  Other = 4,
  DLC = 5,
  SpecialEvents = 6,
  Major_DLC = 7,
  Dungeon5 = 8,
  Dungeon6 = 9,
}

export interface WorshipperSlotAndColor {
  Slot: string;
  color: UnityColor;
}

export interface WorshipperSlotsAndColours {
  SlotAndColours: WorshipperSlotAndColor[];
  AllColor: UnityColor;
}

export interface WorshipperCharacterSkin {
  Skin: string;
}

export interface WorshipperSkinAndData {
  Title: string;
  DropLocation: WorshipperDropLocation;
  Hidden: boolean;
  Invariant: boolean;
  LockColor: boolean;
  TwitchPremium: boolean;
  Skin: WorshipperCharacterSkin[];
  SlotAndColours: WorshipperSlotsAndColours[];
}